Reschedule as soon as we wake.
Prior to this change, if a sync took longer than the wakelock timeout,
background mail checking would not get rescheduled until the phone was
woken for some other reason for a long enough period to complete the
mail check.

With this change, the next sync will immediately be rescheduled, so
even if the check takes too long, at the next check time, the phone
will awake and start work on completing the sync.  This new approach
can lead to redundant "sync" Runnables being enqueued.  However, the
Runnables have a check to make sure they don't do the hard work too
frequently for any given folder, anyway.  If this becomes a problem,
in spite of that check, we could add a way to remove Runnables for the
background queue based on Description.