Improved synchronization timing and WakeLock algorithm.  This new approach should guarantee that the syncing never stalls just because a single attempt exceeds the (currently 10 minute) WakeLock timeout.  At the sync frequency, K-9 will wakeup again and refresh the WakeLock for another go at syncing.  This should make synchronization much more reliable in the face of slow or unavailable connections, in particular ones which hit the socket read timeouts.
Also, fixed the audible notifications so that the ringtone selected
is always for one of the accounts with new unread mail.

Developers now have the option to turn on a fast-flashing LED during
synchronization in order to observe whether synchronization is
happening while the handset is apparently asleep.  I think this option
would be aggravating to non-developer users.  See
Email.NOTIFICATION_LED_WHILE_SYNCING


Issue 175
Issue 170