Add ability to Flag/Unflag messages.  Flagged messages are shown with a red/bold subject.  Can be done in the FolderMessageList using the hotkey G (toggle) or by long-pressing the message to bring up the message context menu.  Can be done while viewing an individual message by using the hotkey G (toggle).  To-do: add a Flag/Unflag option to a menu in the individual message view.
Add sending the Answered flag to the IMAP server when replying to a
message.  This is currently done as soon as the reply is started.
Perhaps it would be better to wait until the reply is sent, but that
would require keeping additional state with the draft.  Messages to
which the user has replied are indicated with a white arrow next to
the subject.

To-do: In the MessageView, provide visual indication of Flagged and
Answered states.