Issue 364
Provide preference for which folder is automatically expanded when
entering the Folder/Message list.

Can be set to -NONE-, so that no folder is automatically expanded.

Defaults to Inbox, which is like the default behavior of the core
Android Email.

Fails to move the automatically opened folder to the top of the
screen.  This needs more work.

The already selected folder should be highlighted (or otherwise
indicated) in the ChooseFolder Activity.  This needs more work.