Issue 4
Operational implementation of Push mail using IMAP IDLE.  Fetches new
mail from the Inbox of any IMAP account which is set with an automatic
sync interval of greater than one minute.

Needs to be enhanced to use a "Folder Push Class" for determining
which folders will participate in Push mail.

Does delete messages or update the flags of messages based on push
notifications.