Issue 1116
Implement batch copy and move, but only for real folders, not search
results.  Search result move/copy is not practical, as we do not
implement cross-account copy, and have no way to choose a target
folder among all of the accounts.

Also, fix the hashCode on LocalStore.LocalFolder so that
MessagingController.actOnMessages will properly aggregate into
Account/Folder lists, even for search results.