Finally get the interaction right between integrated folder, searchable folders, and the various search strategies.
Integration takes precedence over searchability.  The Integrated Inbox
doesn't care about folder searchability.  Nor does an unread or flag
search on specific folders.

Searches on specific accounts always go for displayable folders.  That
might need to change so that searches on specific accounts can be on
DISPLAYABLE or ALL, but never NEVER.  But in the current approach,
account-specific searches are done from the Accounts Activity, which
has an unread counter based on displayable non-special folders, so the
present implementation is the most logical for now.