Issue 258 (e-mail threading support) First pass at attempting to avoid StackOverflowError (SOE) in case of deep threading. SOE can still occur in the "4. Prune empty containers." phase of the threading algorithm.
When it happens, we try to ignore the error and go on processing. Since we're not displaying actual threading (we merely group them), we don't need computed links between existing messages so skipping that phase of the algorithm may be disabled in the future. We'll just have to handle empty containers properly when grouping (already done?).