Don't call super.onBackPressed()
MessageView contained code that would call super.onBackPressed() if the
parent class had such a method (checked using reflection). But the
bytecode verifier prevented loading of MessageView because it contained
a (static) call to super.onBackPressed(). For this to work the method
had to be called using reflection.
I removed this code entirely since we don't do it in MessageList and
FolderList either. And all the time it worked just fine.
With that change we can re-enable support for Android 1.5-2.0.1 devices.