Changed the way we handle menu item clicks using MenuPopupHelper
Overriding onItemClick() doesn't allow us to get the MenuItem that was
clicked when at least one menu item is hidden. This is because the
position in the list of visible menu items is returned. But that value
can't be used to retrieve the corresponding MenuItem (getItem() uses
the position as index into the list of all menu items).