This implements issue 362 (http://code.google.com/p/k9mail/issues/detail?id=362)
Newly-downloaded unread messages are flagged as "unseen", independently of their "unread" state. The count of unseen messages is shown as a separate chip in the folder and account lists (using an "eye" for the chip right now - better ideas welcome). When the folder is opened in the UI, the "unseen" flag is cleared on messages in that folder, as you have now seen them in the folder. If you've used Thunderbird before, this is feature parity with the blue highlighting on folder and account names that it does when new unread messages are downloaded.

A message is considered unseen if k9mail would post a notification about it if the user had notifications enabled. The unseen flag is removed from existing messages if the message is marked as read or deleted from another mail client. Notifications for an account are canceled if the number of unseen messages in that account hits 0.

MAYBE TODO:
- needs a pref to enable/disable it ?