Fix restoring list state if there were content changes since leaving the fragment.
In this case, the list view tries to restore the position valid before the
change by itself. As applying the restored position happens asynchronously,
doing so it overwrites the position we restored. Fix that by enforcing a
position sync right after we restored the position.

As there's no 'enforce sync' API call, this involves a horrible kludge by
calling an unrelated method that just happens to do the sync.