MessageView: fix crash when clicking on URLs in WebView
The ContextThemeWrapper added in a74d57cb71d67e42e1db709e00d5fd514b1ecd0c
used getActivity().getApplicationContext() to get the base context.
This is wrong, because an Application context won't work for starting
activities from the WebView.

Instead, use the context that is given to us in the constructor as base
context. This is the one that would also  be used if no ContextThemeWrapper
was present at all.