Use the KeyChainKeyManager class to select and access creds in the ICS private creds store.
Note: two abuses of statics here:

1) during the 'check settings' activity, we want to prompt the user to
select the client certificates, which requires a reference to that
Activity, so it is temporarily saved in a static variable in the
TrustManagerFactory then cleared.

2) simalarly, once a cert. has been chosen, the alias that was selected
is held in a static String in the TrustManagerFactory to be grabbed
later to use in updating the Account/preferences.