Improve MessageWebView display.
Now that MessageWebView has 'setUseWideViewPort(true)',
the wide view port is excessively wide.  It turns out
Android is using a fixed width of 980 px, so that even
plain text messages (which are already wrapped to fit
the screen) have a large empty area beside them when
scrolled to the left.

Injecting a meta tag in the html header fixes the
problem.