Revert "Fix for StrictMode error."
This reverts commit bbdec62e37210d6abf39cb32a8a6651483d88382.

Aside from being the incorrect solution for fixing the problem
described in pull request 211, the patch generates 'Dead code'
warnings inside the if(){} statements on lines 46 and 47.

The correct fix for the problem was already implemented in commit
5678786c97596703b1b466e7f474f391df617ec1.

Although the logcat in the pull request was generated after the fix,
line numbers in the log indicate that it was based on an outdated
version of MimeUtility.java from before the fix.