Issue 5619: Illegible menu icons in Gingerbread
For Honeycomb and later devices, icons are never displayed
on the overflow menu of the action bar, so it's somewhat
pointless to specify icons for menu items with the attribute
'android:showAsAction="never"' (and "never" is the attribute's
default value if not explicitly set).

The problem is that icons are still being displayed for overflow
menu items on Gingerbread devices when hitting the menu key.
But because the new Holo-themed icons were not designed for
Gingerbread (which  has a different design spec), they tended
to be very difficult to see.

This patch reverts to using Gingerbread icons for menu items with
showAsAction="never".

Ideally, this would be solved in ABS:
https://groups.google.com/forum/?fromgroups#!topic/actionbarsherlock/ce2m7ZHgAJM