Fixes for automated change log publication
The build.xml script was failing if there wasn't already a gh-pages
branch in the local git repo.

Changes:

Fetch the origin/gh-pages branch.

Create a temporary "gh-pages-tmp" branch using origin/gh-pages as the
starting point.  And then as before, copy the changelog to the branch,
push the branch back to origin, and finally delete the temporary branch.

Added more comments to the script.

Ref: comments on Github attached to commit k9mail/k-9@09c27d9.