Remove the added header before sending, to avoid leaking it in the sent message.
Note: AFAICS, the header remains in the outbox message (which is later
moved to the sent folder). This is good I think, because if the message fails to
send, then the correct folder will still be used on a later retry.
I don't think there's a problem with this field remaining in the sent folder.